# Optimize the compute implementation of several operators.