- develop代码：
  ```
  commit 800f5ce6042fb55fafba206a59854bd7c81e324f
  Author: Yiqun Liu <liuyiqun01@baidu.com>
  Date:   Tue Feb 11 15:59:44 2020 +0800

      Optimize the InferShape of several operators. (#2839)
  ```
  - StepRNN的执行时间：**0.0267 ms**
  ```
  I0211 10:17:13.587252  2203 generate_program_pass.h:37] insts.size 42
  I0211 10:17:13.865773  2203 test_step_rnn_lite_x86.cc:79] warmup: 10, repeats: 10000, spend 0.0266628 ms in average.
  ```
- 优化Compute的PR：[Paddle-Lite#2843](https://github.com/PaddlePaddle/Paddle-Lite/pull/2843)
  - StepRNN的执行时间：**0.0217 ms**
  ```
  I0211 09:19:36.004922 13240 generate_program_pass.h:37] insts.size 42
  I0211 09:19:36.234015 13240 test_step_rnn_lite_x86.cc:79] warmup: 10, repeats: 10000, spend 0.0217213 ms in average.
  ```

- gperftools分析各个函数的耗时占比
  - 通用优化方法：
    - 方法一，输入Tensor的dims的使用改成引用类型，非引用类型将发生一次DDim的拷贝。
    - 方法二，避免反复调用push_back、emplace_back函数，最好先计算出元素个数，通过resize接口一次申请好空间。
    - 方法三，Tensor转`EigenVector`优化。softmax、reduce_sum、gru、softsign均可受益。

| module | e73d46 | #2839 (develop) | #2843 | 优化点介绍 |
|---|---|---|---|---|
| Activation::InferShape | 0.99% | 1.83% | 1.55% |
| ConcatOpLite::InferShape | 6.93% | 2.44% | 2.33% |
| FcOpLite::InferShape | 0.99% | 2.44% | 3.88% |
| GRUOpLite::InferShape | 1.49% | 1.22% | 0.78% |
| LookupTableOpLite::InferShape | 18.32% | 4.27% | 9.30% |
| ReduceOp::InferShape | 3.96% | 3.05% | 0.78% |
| Reshape2Op::InferShape | 5.45% | 5.49% | 3.88% |
| SequenceReshapeOp::InferShape | 0.99% | | 2.33% |
| KernelBase::Launch | 57.92% | 77.44% | 73.64% |
| ConcatCompute::Run | 0.99% | 3.05% | 2.33% | 方法一<br>将只有一个输入Tensor的if判断提前，可以提前返回，避免用不到的操作。
| FcCompute::Run | 9.41% | 13.41% | 16.28% |
| GRUCompute::Run | 20.79% | 26.22% | 18.60% | 方法一 + 方法二；<br>直接调用`mutable_lod`获取lod的指针，而不是先定义一个LoD用于计算，可避免一次LoD的拷贝。
| LookupTableCompute::Run | 2.48% | 4.27% | 5.43% |
| ReduceSumCompute::Run | 4.95% | 17.07% | 15.50% | 方法一<br>输出Tensor的dims已经在`InferShape`中计算好，没必要在Compute里面重复计算。
| Reshape2Compute::Run | 2.48% | 1.83% |
| SequenceReshapeCompute::Run | 1.98% | 3.66% | 3.88% | 方法一<br>直接用正确的dims来`Resize`输出Tensor。
| SoftmaxCompute::Run | 8.91% | 3.66% | 5.43% | 方法一
| SoftsignCompute::Run | 3.96% | 3.05% | 3.10% |

**备注**：此PR由于单测已过，本地分支merge了develop代码之后（没有push到github）使用gperftools分析各个函数的耗时占比和StepRNN的整体时间。